applications/vehicle-pos-data-forwarder-python/requirements.txt
confluent-kafka==1.5
paho-mqtt==1.5